Stop planning according to consequences, not just order
This will make sure we don’t downplay actions that lead to a lot of ‘pass’/‘regain balance’ turns where there is not much happening. These consequences ‘eat’ order without much data.

For example, if a move of order n ’forces’ the opponent to do just one thing, there will be only one consequence in order n+1. The benefit of that move could materialize in n+3 but we might never get there.