Generate self getter for actions
To access methods like `movePlayer`, SimpleActions need a reference to themselves, which is given as self. On the other hand, normal RoamingAction subclasses don’t need that (they can access `movePlayer` directly, as it’s their method.

The problem is that from the writer’s perspective, it’s unclear whether to use `movePlayer` or `self.movePlayer`. So we add a convenience getter `self` and unify this.