Fix bug with pronouns
When there was something like `<subject> {kill<s> <object>|chop<s> off <object’s> head}`, we made the `<object’s>` into `his`. So sometimes we got reports that were referring to people by pronouns even when it was confusing.

The problem was that the `<object’s>`-to-pronoun logic was applied _before_ Randomly parsed the string. This meant that the `<object’s>` was made into a pronoun as if it was in the same sentence as the alternative sentence.

Fixed by resolving Randomly as soon as we get text from the report.