Make attack situations DRY
Instead of reimplementing attack situations from scratch, use a single AttackSituation class. Now that top-level functions can be persisted, it’s possible to do this.

This is a chore — but there are many attack situations and all of them are tens of lines long. This change compacts them into 2-3 lines.