Simplify API of visitHistory
Instead of forwarding query calls from WorldState, just call the query on the member itself.

We still need the forwarding call for `addVisit` because that’s making a complicated call on a `VisitHistoryBuilder` instance, and we can’t make that easier on `VisitHistoryBuilder` itself since it’s generated from source. The only option would be to create a custom builder, but that’s a stretch.