Add StatefulRandom to WorldState
Fixes https://trello.com/c/OxEtmmHw/599-predictable-randomness. Most importantly, it introduces `.randomInt()`, `.randomDouble()`, and `.randomBool()`,  into `WorldStateBuilder`, so that we can use them in actions.

This allows actions (more precisely, their apply functions) to have random outcomes, while this randomness is predictably a wholy defined by the current state of the world.

Real randomness is still allowed, but only for things like NLG. The state mustn’t depend on randomness.

Every entity must now instantiate with an `id`. We stop using `UniqueIdMaker` because of a potential bug with duplicite IDs (https://trello.com/c/OxEtmmHw/599-predictable-randomness#comment-5ad6c52dedce2b4ff72973c9). Instead, we use `StatefulRandom` to create ids.

We’re now also allowing to inject (seeded) Random into `CliRunner`, which allows us to not only have a predictable randomness in the Book, but also in the choices. This is used to test that two walkthroughs with the same seed end up completely identical.

We’re using `WorldState.statefulRandomState` also in `Planner`, so that NPCs predictably choose actions. This is a bit shaky, as that’s using the random state outside an action, and therefore it doesn’t update the `WorldState.statefulRandomState` variable.