Don’t spend time planning ahead for the player
We are showing all options for the player anyway these days, so planning for the player is a waste of resources.

Formerly, we only showed a subset of actions to the player, and for that, we had to rank them so that only the best options would be presented. But now that we’re thinking of using JRPG-style mutli-level menus, there is no need for pruning options for the player.