Start using Action.commandPath
`Action.commandPath` takes the “ >> “ approach and makes it official. All actions can now be shown to the player in a tree of submenus.

We are still keeping `Action.command` so that an action can have both a literary command (“Kick the goblin to the ground”) and a sequence of UI commands (“Attack goblin > Stance > Kick”).