Exercise ChoiceTree parsing
In order to catch bugs like the one addressed with 4797327c `Fix crash when two choices are identical` earlier, we are now parsing every ChoiceBlock with ChoiceTree even in CliRunner. CliRunner doesn’t use the tree (yet?) but it’s used by fuzzy test runner, so crashing bugs should be caught much earlier.

This also changes the logic in CliRunner. It is assumed that implicit choice is not the one in `actionPattern`, and so implicit choices are always just picked automatically.