Fix <object> and <object2> without entities
Until now, it was possible to have a storyline like `<subject> fend<s> off the slash{|with <object2>}`. This meant that sometimes, a report would appear without the object2 mentioned. Then the next sentence would be rendered as if object2 _was_ mentioned (because it was non-null in the Report).

Also adds MoveEntity to AttackerSituation. This allows storyline to consistently refer to the maneuver (slash, feint, etc.) without creating a new entity every time.