Use initiative to set up who starts in each fight
We were ignoring initiative until now.

This change also makes SlayMonsters non-proactive. This is because we set up Actor.recoveringUntil in SlayMonsters, and we don’t want the default behavior which modifies recoveringUntil for the actor performing each proactive action. It also makes more sense, since SlayMonsters is an implicit action.