Make invincible actors truly invincible
So far, they could get to `hitpoints == 0` but would be ressurected after the fight. This worked for friendlies, but not for enemies.

Now, it’s not possible to deal the final blow that would make an invincible actor dead. This helps in the first fight, where we make sure that the goblin doesn’t get killed by Tamara, by sheer luck, in the first two turns. After that, all bets are off.