Generalize “<subject>” / “<object>” etc. stopwords
This introduces ComplementType, which is a struct with stopwords for a particular grammar argument (such as subject or “objectOwner”).

The change also removes the “feature” which ignored missing owners. Owners and object-owners now mustn’t be null.

owners and objectOwners are now treated the same as subject, object and object2.