Remove theOtherNoun identifier
We are not using it, and it meant that some problems were encountered later than they should have.

For example, when an entity lost all identifiers in `_detectFirstMentions`, we wouldn’t know about it in the following `__assertAtLeastOneIdentifier`, because `theOtherNoun` was still lingering (so the set wasn’t, technically, empty).