Fix bug when else block is missing in [[IF]]
This didn’t work, because there was no `?` after the ELSE block in the Regular Expression (so it was always required).

Additionally, a misunderstanding of the API: I thought groupCount would count less groups if some of them didn’t match this time (as in, a `?` group). But the groupCount is probably always the same. Instead, groups can return `null`.