Implement context helpers as extension methods
Instead of calling `$(c).something`, we can now do `c.something`. And it’s statically typed.

This also adds `c.inRoomParent`.

Also, there’s a semantic change of FOR_LOCATION. An action is considered available for a location if it fits the _parent_ name of the location. So, all variants of an action will get it. This also means that if you define an action for a VARIANT, then it will never be available.

It also lifts the SDK version to 2.7.0 (where static extension methods are stable).