Catch errors via Zone
This allows for much better debugging. Using the previous `try-catch` approach, when anything in the game threw, we would not get access to the deeper frames. In other words, we would only see the error object, and the `StackTrace` object. Catching the error via async try catch would not allow the developer to inspect where the error was thrown, nor the state of variables at that point.

Simply going with `runZoned` means that the debugger stops where it should, and the developer can inspect the state of the variables at every frame.