Fix bug with second report
Fixes https://trello.com/c/bsF5aJUi/1148-bug-second-orc-not-made-into-pronoun.

The issue was that we added all the storyline elements into the first report of the storyline. This led to the _second_ report in the storyline to try and avoid confusing identifiers. This led to things like "I dodge the orc and the orc falls to the muddy ground.” (note the second orc, which could easily be “he”).

This is solved by adding all storyline entities into a “bogus” -1 (minus first? zero-th?) report. Since the `_getIdentifiersThroughoutStory` logic works with the concept of a “previous” report, this works perfectly. It’s as if the NLG system is “thinking” a zero-th report with all the entities in it, so that the first report doesn’t start with “orc” when there are several orcs present, but it doesn’t actually render it.