Add `isRaw` to report
So far, we’ve been abusing `Report.wholeSentence` for writer’s raw text (i.e. not to be modified by the NLG).

But that doesn’t work, because sometimes `wholeSentence` actually needs to be processed by the NLG, including capitalization. The real meaning of `wholeSentence` is: don’t put this as part of a larger sentence (but otherwise, do your NLG stuff).

I’ve added `Report.isRaw` for things where the NLG shouldn’t touch the report at all.