Disallow caravan and offensive when in populated rooms
By themselves, there’s no reason to stipulate that these events cannot happen when the player is in a populated room, like the Staging Area.

But, since — for example — the arrival of the caravan will end the Quake 1 variant of the Staging Area, we’d effectively change the room from under the player.

To stay conservative, we just don’t allow changing rooms like that.