Implement override actor describers
By default, actors in a room are described with something like “This and this is here.” But oftentimes, we want to say this in the description prose instead, or we want the actor to do something (like wave at the player).

Now we have a map that does just that.