Fix bug with false positive situation.monstersAlive
We assumed the player will always be the one recording the `SlayRecord`. But it’s perfectly possible for Tamara, for example, to make the final blow. Now, the `SlayRecord` is under her ID, and so when we re-enter the room, if we ask for only slayings by the player, none will show up.

Fixed by removing the question about the actor of the slaying altogether.