Respect Approach.isApplicable for “fast travel” routes
Before this change, only the last “isApplicable” was heeded. Now, any failing isApplicable check along the way will cancel the route. (Which is what you want. If there’s a rock blocking the way on the way to an exit, then we don’t want the algorithm to give us ways that go through that rock.