Fix bug which led to “His sword cuts his neck”
This was a problem with a false negative when checking if a second iteration of ShadowGraph should be run.

We added a new entity to one of the reports in the Storyline (a sword’s owner, to differentiate it from another sword). But the same entity happened to be mentioned elsewhere in the Storyline, so `_entitiesHadToBeAdded` was set to `false`, and the result accepted as the final ShadowGraph.

The fix is that we force the ShadowGraph construction to happen at least twice, and directly compare the entities for each report. Only if two consecutive runs have exactly the same entities do we approve the ShadowGraph.