Avoid ShadowGraph failure with severed body parts
In a Storyline invovling a severing of an arm, for examlpe, there could be two entities:

- The arm, as it was attached to the body (a BodyPart)
- The arm, after it was severed (an Item).

This produced an error, since the two entities have a different ID and the same name. One of them didn’t have an adjective and the other one did (“severed”). That way, NLG couldn’t distinguish between the two.

This is fixed by actually changing the name of the body part (a “right hand” body part becomes a “severed right hand” item).