Avoid hateTowards bug
When there was Confusion in any actor’s future, it meant that the 1000 score totally skewed all results.

The issue was that hateTowards was calculated from the perspective of the _future_ actor. For example, if there was a future in which a goblin gets confused, suddenly that consequence means the enemy team score jumps up by thousands (because suddenly all living friends are hated-towards by a 1000 intensity.

This is fixed by using the initial actor (confused or not) to score the future. This makes more sense even by intuition. We don’t value future possibilities by how our future selves will like them but how we like them today. Our future self might be insane and love complete solitude, but that shouldn’t make us motivated to get insane and into the wilderness.