Make all StatUpdate and StatInitialization int
Before this change, stats were generic. But this posits a big problem when deserializing on the other side (it would deserialize to StatUpdate<Object>).

Since we’re now (de)serializing between the worker (book) thread and the main (presenter) thread, we need to do this.
See flutter_egamebook 0b718ab9 Serialize elements before sending them to the main thread.

If we ever need to send other stats than int, we’ll have to create a separate class (like DateTimeStatUpdate).