Fix a gnarly bug with player
When `c.player` was defined as a method extension, there was a gnarly bug. The `player` getter was both on `ActionContext` and on `ApplicabilityContext`. When using `c.hasItem`, we were using original world unwittingly even when calling it on `ActionContext`, because `hasItem()` was defined on `ApplicabilityContext`.

We are fixing it by moving `.player` from extensions to ActionContext and ApplicabilityContext itself. That way, ActionContext will always get the most current `.player`.